update test suite for status/update
