+++ src/modules/statusFunctions.js
@@ -7,4 +7,3 @@ export function updateTaskStatus(tasks, index, completed) {
 export function clearCompletedTasks(tasks) {
   return tasks.filter((task) => !task.completed);
 }
-
